updated database and added nodemailer

+++ src/entity/Item.ts
@@ -7,7 +7,7 @@ export class Item{
     @PrimaryGeneratedColumn()
     id: number;
 
-    @Column("decimal")
+    @Column("integer")
     price: number;
 
     @Column("text")
